requirements.txt
Django==3.1.1
djangorestframework==3.11.1
psycopg2==2.8.6

